Extract Ivy specific tests into their own test case
