+++ subprojects/dependency-management/src/integTest/groovy/org/gradle/integtests/resolve/ComponentMetadataRulesIntegrationTest.groovy
@@ -15,14 +15,11 @@
  */
 package org.gradle.integtests.resolve
 
-import org.gradle.api.internal.artifacts.ivyservice.NamespaceId
 import org.gradle.integtests.fixtures.GradleMetadataResolveRunner
 import org.gradle.integtests.fixtures.RequiredFeature
 import org.gradle.integtests.fixtures.RequiredFeatures
-import org.gradle.test.fixtures.encoding.Identifier
-import spock.lang.Unroll
 
-class ComponentMetadataRulesIntegrationTest extends AbstractModuleDependencyResolveTest {
+class ComponentMetadataRulesIntegrationTest extends AbstractModuleDependencyResolveTest implements ComponentMetadataRulesSupport {
     String getDefaultStatus() {
         GradleMetadataResolveRunner.useIvy()?'integration':'release'
     }
@@ -258,334 +255,6 @@ dependencies {
         succeeds 'resolve'
     }
 
-    // Ivy specific tests
-    @RequiredFeatures(
-        @RequiredFeature(feature = GradleMetadataResolveRunner.REPOSITORY_TYPE, value = "ivy")
-    )
-    def "can access Ivy metadata by accepting parameter of type IvyModuleDescriptor"() {
-        given:
-        repository {
-            'org.test:projectA:1.0' {
-                withModule {
-                    withExtraInfo((ns('foo')): "fooValue", (ns('bar')): "barValue")
-                    withBranch('someBranch')
-                    withStatus('release')
-                }
-            }
-        }
-
-        buildFile <<
-            """
-def ruleInvoked = false
-
-dependencies {
-    components {
-        all { ComponentMetadataDetails details, IvyModuleDescriptor descriptor ->
-            ruleInvoked = true
-            assert descriptor.extraInfo.asMap() == [${declareNS('foo')}: "fooValue", ${declareNS('bar')}: "barValue"]
-            assert descriptor.extraInfo.get('foo') == 'fooValue'
-            assert descriptor.extraInfo.get('${ns('foo').namespace}', 'foo') == 'fooValue'
-            assert descriptor.branch == 'someBranch'
-            assert descriptor.ivyStatus == 'release'
-        }
-    }
-}
-
-resolve.doLast { assert ruleInvoked }
-"""
-        when:
-        repositoryInteractions {
-            'org.test:projectA:1.0' {
-                expectResolve()
-            }
-        }
-
-        then:
-        succeeds 'resolve'
-        // also works when already cached
-        succeeds 'resolve'
-    }
-
-    @RequiredFeatures(
-        @RequiredFeature(feature = GradleMetadataResolveRunner.REPOSITORY_TYPE, value = "ivy")
-    )
-    def "produces sensible error when accessing non-unique extra info element name"() {
-        given:
-        repository {
-            'org.test:projectA:1.0' {
-                withModule {
-                    withExtraInfo((ns('foo')): "fooValue", (new NamespaceId('http://some.other.ns', 'foo')): "barValue")
-                }
-            }
-        }
-
-        buildFile <<
-            """
-def ruleInvoked = false
-
-dependencies {
-    components {
-        all { ComponentMetadataDetails details, IvyModuleDescriptor descriptor ->
-            ruleInvoked = true
-            descriptor.extraInfo.get('foo')
-        }
-    }
-}
-
-resolve.doLast { assert ruleInvoked }
-"""
-
-        and:
-        repositoryInteractions {
-            'org.test:projectA:1.0' {
-                expectGetMetadata()
-            }
-        }
-
-        when:
-        fails 'resolve'
-
-        then:
-        failure.assertHasDescription("Execution failed for task ':resolve'.")
-        failure.assertHasLineNumber(48)
-        failure.assertHasCause("Could not resolve all files for configuration ':conf'.")
-        failure.assertHasCause("Could not resolve org.test:projectA:1.0.")
-        failure.assertHasCause("Cannot get extra info element named 'foo' by name since elements with this name were found from multiple namespaces (http://my.extra.info/foo, http://some.other.ns).  Use get(String namespace, String name) instead.")
-    }
-
-    @Unroll
-    def "can access Ivy metadata with #identifier characters"() {
-        given:
-        def branch = identifier.safeForBranch().decorate("branch")
-        def status = identifier.safeForFileName().decorate("status")
-        repository {
-            'org.test:projectA:1.0' {
-                withModule {
-                    withBranch(branch)
-                    withStatus(status)
-                }
-            }
-        }
-
-        buildFile <<
-            """
-def ruleInvoked = false
-
-dependencies {
-    components {
-        all { details, IvyModuleDescriptor descriptor ->
-            ruleInvoked = true
-            assert descriptor.branch == '${sq(branch)}'
-            details.statusScheme = [ '${sq(status)}' ]
-            assert descriptor.ivyStatus == '${sq(status)}'
-        }
-    }
-}
-
-resolve.doLast { assert ruleInvoked }
-"""
-
-        expect:
-        // Cannot use @RequiredFeatures because of Unroll above
-        if (GradleMetadataResolveRunner.useIvy()) {
-            repositoryInteractions {
-                'org.test:projectA:1.0' {
-                    expectResolve()
-                }
-            }
-            succeeds 'resolve'
-        }
-
-        where:
-        identifier << Identifier.all
-    }
-
-    @RequiredFeatures(
-        @RequiredFeature(feature = GradleMetadataResolveRunner.REPOSITORY_TYPE, value = "ivy")
-    )
-    def "rule that doesn't initially access Ivy metadata can be changed to get access at any time"() {
-        given:
-        repository {
-            'org.test:projectA:1.0' {
-                withModule {
-                    withExtraInfo((ns('foo')): "fooValue", (ns('bar')): "barValue")
-                    withBranch("someBranch")
-                    withStatus("release")
-                }
-            }
-        }
-
-        def baseScript = buildFile.text
-
-        when:
-        repositoryInteractions {
-            'org.test:projectA:1.0' {
-                expectResolve()
-            }
-        }
-        buildFile.text = baseScript +
-            """
-def ruleInvoked = false
-
-dependencies {
-    components {
-        all { ComponentMetadataDetails details ->
-            ruleInvoked = true
-        }
-    }
-}
-
-resolve.doLast { assert ruleInvoked }
-"""
-
-        then:
-        succeeds 'resolve'
-
-        when:
-        buildFile.text = baseScript +
-            """
-def ruleInvoked = false
-
-dependencies {
-    components {
-        all { ComponentMetadataDetails details, IvyModuleDescriptor descriptor ->
-            ruleInvoked = true
-            assert descriptor.extraInfo.asMap() == [${declareNS('foo')}: "fooValue", ${declareNS('bar')}: "barValue"]
-            assert descriptor.branch == 'someBranch'
-            assert descriptor.ivyStatus == 'release'
-        }
-    }
-}
-
-resolve.doLast { assert ruleInvoked }
-"""
-
-        then:
-        succeeds 'resolve'
-    }
-
-    @RequiredFeatures(
-        @RequiredFeature(feature = GradleMetadataResolveRunner.REPOSITORY_TYPE, value = "ivy")
-    )
-    def "changed Ivy metadata becomes visible once module is refreshed"() {
-        def baseScript = buildFile.text
-
-        given:
-        repository {
-            'org.test:projectA:1.0' {
-                withModule {
-                    withExtraInfo((ns('foo')): "fooValue", (ns('bar')): "barValue")
-                    withBranch('someBranch')
-                    withStatus('release')
-                }
-            }
-        }
-
-        buildFile.text = baseScript +
-            """
-def ruleInvoked = false
-
-dependencies {
-    components {
-        all { ComponentMetadataDetails details, IvyModuleDescriptor descriptor ->
-            ruleInvoked = true
-            assert descriptor.extraInfo.asMap() == [${declareNS('foo')}: "fooValue", ${declareNS('bar')}: "barValue"]
-            assert descriptor.branch == 'someBranch'
-            assert descriptor.ivyStatus == 'release'
-        }
-    }
-}
-
-resolve.doLast { assert ruleInvoked }
-"""
-
-        when:
-        repositoryInteractions {
-            'org.test:projectA:1.0' {
-                expectResolve()
-            }
-        }
-
-        then:
-        succeeds 'resolve', '--refresh-dependencies'
-
-        when:
-        resetExpectations()
-        repository {
-            'org.test:projectA:1.0' {
-                withModule {
-                    withExtraInfo((ns('foo')): "fooValueChanged", (ns('bar')): "barValueChanged")
-                    withBranch('differentBranch')
-                    withStatus('milestone')
-                    publishWithChangedContent()
-                }
-            }
-        }
-
-        then:
-        succeeds 'resolve'
-
-        when:
-        buildFile.text = baseScript +
-            """
-def ruleInvoked = false
-
-dependencies {
-    components {
-        all { ComponentMetadataDetails details, IvyModuleDescriptor descriptor ->
-            ruleInvoked = true
-            file("metadata").delete()
-            file("metadata") << descriptor.extraInfo.asMap().toString()
-            file("metadata") << "\\n"
-            file("metadata") << descriptor.branch
-            file("metadata") << "\\n"
-            file("metadata") << descriptor.ivyStatus
-        }
-    }
-}
-
-resolve.doLast { assert ruleInvoked }
-"""
-
-        and:
-        resetExpectations()
-        repository {
-            'org.test:projectA:1.0' {
-                withModule {
-                    withExtraInfo((ns('foo')): "fooValueChanged", (ns('bar')): "barValueChanged")
-                    withBranch('differentBranch')
-                    withStatus('milestone')
-                    publishWithChangedContent()
-                }
-            }
-        }
-        repositoryInteractions {
-            'org.test:projectA:1.0' {
-                expectHeadMetadata()
-                withModule {
-                    // todo: handle this properly in ModuleVersionSpec test fixture
-                    getArtifact(name: 'ivy', ext: 'xml.sha1').allowGetOrHead()
-                    if (GradleMetadataResolveRunner.isGradleMetadataEnabled()) {
-                        getArtifact(ext: 'module.sha1').allowGetOrHead()
-                    }
-                }
-                expectGetMetadata()
-                expectHeadArtifact()
-                expectGetArtifact()
-                withModule {
-                    getArtifact(ext: 'jar.sha1').allowGetOrHead()
-                }
-            }
-        }
-
-        then:
-        succeeds 'resolve', '--refresh-dependencies'
-        assert file("metadata").text == "{{http://my.extra.info/bar}bar=barValueChanged, {http://my.extra.info/foo}foo=fooValueChanged}\ndifferentBranch\nmilestone"
-    }
-
-    @RequiredFeatures(
-        @RequiredFeature(feature = GradleMetadataResolveRunner.REPOSITORY_TYPE, value = "ivy")
-    )
     def "produces sensible error when @Mutate method does not have ComponentMetadata as first parameter"() {
         buildFile << """
             dependencies {
@@ -653,19 +322,4 @@ resolve.doLast {
         succeeds 'resolve'
     }
 
-    def ns(String name) {
-        return new NamespaceId("http://my.extra.info/${name}", name)
-    }
-
-    def declareNS(String name) {
-        return "(new javax.xml.namespace.QName('http://my.extra.info/${name}', '${name}'))"
-    }
-
-    String sq(String input) {
-        return escapeForSingleQuoting(input)
-    }
-
-    String escapeForSingleQuoting(String input) {
-        return input.replace('\\', '\\\\').replace('\'', '\\\'')
-    }
 }
